requirements.txt
alembic
blinker
click
colorama
Flask
flask-cors
Flask-JWT-Extended
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
GeoAlchemy2
greenlet
gunicorn
itsdangerous
Jinja2
Mako
MarkupSafe
packaging
psycopg2-binary
PyJWT
python-dotenv
SQLAlchemy
typing_extensions
Werkzeug
Flask-Mail
flask-socketio
simple-websocket
gevent
reportlab
Flask-APScheduler